+++ Map.py
@@ -1,7 +1,7 @@
 # Course: CS 30
 # Period: 3
 # Date created: 20/9/2021
-# Date last modified: 10/20/2021
+# Date last modified: 10/25/2021
 # Name: Ben Oblinski
 # Description: map system for the game menu
 from numpy import*